+++ travis.script.sh
@@ -372,6 +372,19 @@ else
           's#url: ".*"#url: "https://gchq.github.io/stroom/swagger.json"#g' \
           "${ghPagesDir}/index.html"
     fi
+
+    # If it is a tagged build copy the docker config files with new names so we
+    # can add them as release artefacts.
+    # This is so the stack build can download them
+    if [ -n "$TRAVIS_TAG" ]; then
+      cp \
+        "${TRAVIS_BUILD_DIR}/stroom-app/docker/build/config.yml" \
+        "${TRAVIS_BUILD_DIR}/stroom-app-config-${TRAVIS_TAG}.yml"
+
+      cp \
+        "${TRAVIS_BUILD_DIR}/stroom-proxy/stroom-proxy-app/docker/build/config.yml" \
+        "${TRAVIS_BUILD_DIR}/stroom-proxy-config-${TRAVIS_TAG}.yml"
+    fi
 fi
 
 exit 0